Update typehinting, fix imports, correct description
This function now has proper typehinting, and all trailing whitespaces have been erased.

I've also moved all native Python imports outside of the function, since those should come pre-installed with Python anyway.

vsutil and Vapoursynth imports were also added to the top, as those are hard requirements (or in the case of vsutil, should ideally be a hard requirement) for most functions in a *func.

The description has also been updated to properly reflect what is and isn't an optional argument, as I'm not entirely sure why just about everything is marked as optional. Is it just because the user doesn't _need_ to fill those in or something?